requirements.txt
annotated-types==0.6.0
asgiref==3.8.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
coloredlogs==15.0.1
contourpy==1.2.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.12.1
Django==5.0.3
django-debug-toolbar==4.3.0
django-rest-swagger==2.2.0
djangorestframework==3.15.0
drf-spectacular==0.27.1
drf-yasg==1.21.7
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.50.0
fsspec==2024.3.1
ftfy==6.2.0
humanfriendly==10.0
idna==3.6
imageio==2.34.0
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-path==0.3.2
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy-object-proxy==1.10.0
lazy_loader==0.3
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.3
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.17.1
openapi-codec==1.3.2
openapi-core==0.19.0
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.1
parse==1.20.1
pathable==0.4.3
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
protobuf==5.26.0
psutil==5.9.8
py-cpuinfo==9.0.0
pydantic==2.6.4
pydantic_core==2.16.3
PyMatting==1.1.12
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.31.1
regex==2023.12.25
rembg==2.0.56
requests==2.31.0
rfc3339-validator==0.1.4
rpds-py==0.18.0
scikit-image==0.22.0
scipy==1.12.0
seaborn==0.13.2
setuptools==69.2.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
sympy==1.12
thop==0.1.1.post2209072238
tifffile==2024.2.12
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.31
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
